refactor(server): extract operation routes
